+++ packages/backend/.eslintrc.js
@@ -34,7 +34,7 @@ module.exports = {
         },
 
         {
-            // Only throw warning for existing files, new files should be error
+            // Do not check this rule for existing files, new files should be error
             files: [
                 'src/*.ts',
                 'src/analytics/**/*.ts', // TODO fix these folders
@@ -52,9 +52,9 @@ module.exports = {
                 'src/projectAdapters/**/*.ts',
             ], 
             rules: {
-                "@typescript-eslint/no-unsafe-member-access": "warn",
-                "@typescript-eslint/no-unsafe-assignment": "warn",
-                "@typescript-eslint/no-unsafe-call": "warn"
+                "@typescript-eslint/no-unsafe-member-access": "off",
+                "@typescript-eslint/no-unsafe-assignment": "off",
+                "@typescript-eslint/no-unsafe-call": "off"
             }
         },
         {
@@ -63,8 +63,6 @@ module.exports = {
                 '*.mock.ts',
                 '*.test.ts',
                 '*.spec.ts',
-
-
             ],
             rules: {
                 "@typescript-eslint/no-unsafe-member-access": "off",