+++ packages/backend/src/config/lightdashConfig.mock.ts
@@ -94,7 +94,7 @@ export const lightdashConfigMock: LightdashConfig = {
     logging: {
         level: 'debug',
         format: 'pretty',
-        outputs: [],
+        outputs: ['console'],
         consoleFormat: undefined,
         consoleLevel: undefined,
         fileFormat: undefined,