+++ packages/frontend/src/features/omnibar/components/Omnibar.tsx
@@ -177,7 +177,6 @@ const Omnibar: FC<Props> = ({ projectUuid }) => {
                   )
             : [];
     }, [searchResults]);
-
     useEffect(() => {
         setFocusedItemIndex(undefined);
     }, [query, searchFilters]);